Adding type annotations

+++ src/components/Grid.tsx
@@ -1,6 +1,6 @@
 import React, {useEffect, useRef, useState} from 'react';
 import {useTheme} from "../Theme";
-import Table from "./Table";
+import Table, {TableHeaderProp} from "./Table";
 import Gallery from "./Gallery";
 import Card from "./Card";
 import db from "../libs/database";
@@ -11,269 +11,100 @@ import {converter} from "../libs/converter";
 import ComponentEnhancer, {extractComponentProps} from "./ComponentEnhancer";
 import setLog from "../libs/log";
 
+type Column = TableHeaderProp & {
+    onDisplay?: (value: string, record: any, key: string) => string;
+};
 
-function Grid({
-    columns = null,
-    format = {},
-    dataStoragePath = null,
-    dataArray = null,
-    Header = null,
-    HeaderAction = null,
-    Footer = null,
-    allowedActions = null,
-    modalSize = "lg",
-    setModalHeader = null,
-    setPrimaryKey = null,
-    onLoadRecord = null,
-    onDisplayBefore = null,
-    onInsertRecord = null,
-    onUpdateRecord = null,
-    onDeleteRecord = null,
-    onFinallyRecord = null,
-    onClick = null,
-    Form = null,
-    wrapClass = null,
-    scroll = false,
-    type = "table",
-    showLoader =  false,
-    groupBy = null,
-    sticky= null,
-    log= false
-}: {
-    columns?: Array<{ label: string, key: string, sort?: boolean, onDisplay?: (value: string, record: any, key: string) => string }>,
-    format?: { [key: string]: string | ((data: string) => string) },
-    dataStoragePath?: string,
-    dataArray?: Array<{ [key: string]: any }>,
-    Header?: string | React.ReactNode,
-    HeaderAction?: string | React.ReactNode,
-    Footer?: string | React.ReactNode,
-    allowedActions ?: Array<"add" | "edit" | "delete">,
-    modalSize ?: string,
-    setModalHeader ?: (record ?: any) => string,
-    setPrimaryKey ?: (record: any) => string,
-    onLoadRecord?: (record: any, index, number) => any,
-    onDisplayBefore?: (records: any, setRecords: (records: any) => void, setLoader: (loader: boolean) => void) => void,
-    onInsertRecord?: (record: any, key: string) => any,
-    onUpdateRecord?: (record: any, key: string) => any,
-    onDeleteRecord?: (record: any, key: string) => any,
-    onFinallyRecord?: (action: string, record: any, key: string) => any,
-    onClick?: (record: any, key: string) => any,
-    Form?: any,
-    wrapClass?: string,
-    scroll?: boolean,
-    type?: "table" | "gallery",
-    showLoader?: boolean,
-    groupBy?: string | string[],
-    sticky?: "top" | "bottom",
-    log?: boolean
-}) {
-    return (dataArray
-        ? <GridArray
-            columns={columns}
-            format={format}
-            dataArray={dataArray}
-            Header={Header}
-            HeaderAction={HeaderAction}
-            Footer={Footer}
-            allowedActions={allowedActions}
-            modalSize={modalSize}
-            setModalHeader={setModalHeader}
-            setPrimaryKey={setPrimaryKey}
-            onLoadRecord={onLoadRecord}
-            onDisplayBefore={onDisplayBefore}
-            onInsertRecord={onInsertRecord}
-            onUpdateRecord={onUpdateRecord}
-            onDeleteRecord={onDeleteRecord}
-            onFinallyRecord={onFinallyRecord}
-            onClick={onClick}
-            Form={Form}
-            wrapClass={wrapClass}
-            scroll={scroll}
-            type={type}
-            showLoader={showLoader}
-            groupBy={groupBy}
-            sticky={sticky}
-            log={log}
-        />
-        : <GridDatabase
-            columns={columns}
-            format={format}
-            dataStoragePath={dataStoragePath}
-            Header={Header}
-            HeaderAction={HeaderAction}
-            Footer={Footer}
-            allowedActions={allowedActions}
-            modalSize={modalSize}
-            setModalHeader={setModalHeader}
-            setPrimaryKey={setPrimaryKey}
-            onLoadRecord={onLoadRecord}
-            onDisplayBefore={onDisplayBefore}
-            onInsertRecord={onInsertRecord}
-            onUpdateRecord={onUpdateRecord}
-            onDeleteRecord={onDeleteRecord}
-            onFinallyRecord={onFinallyRecord}
-            onClick={onClick}
-            Form={Form}
-            wrapClass={wrapClass}
-            scroll={scroll}
-            type={type}
-            showLoader={showLoader}
-            groupBy={groupBy}
-            sticky={sticky}
-            log={log}
-        />
-    );
+type GridProps = {
+    columns?: Column[];
+    format?: { [key: string]: string | ((data: string) => string) };
+    dataStoragePath?: string;
+    dataArray?: Array<{ [key: string]: any }>;
+    Header?: string | React.ReactNode;
+    HeaderAction?: string | React.ReactNode;
+    Footer?: string | React.ReactNode;
+    allowedActions?: Array<"add" | "edit" | "delete">;
+    modalSize?: string;
+    setModalHeader?: (record?: any) => string;
+    setPrimaryKey?: (record: any) => string;
+    onLoadRecord?: (record: any, index: string | number) => any;
+    onDisplayBefore?: (records: any, setRecords: (records: any) => void, setLoader: (loader: boolean) => void) => void;
+    onInsertRecord?: (record: any, key: string) => any;
+    onUpdateRecord?: (record: any, key: string) => any;
+    onDeleteRecord?: (record: any, key: string) => boolean;
+    onFinallyRecord?: (action: string, record: any, key: string) => any;
+    onClick?: (record: any, key: string) => any;
+    Form?: any;
+    scroll?: boolean;
+    type?: "table" | "gallery";
+    showLoader?: boolean;
+    groupBy?: string | string[];
+    sticky?: "top" | "bottom";
+    log?: boolean;
+    wrapClass?: string;
+};
+
+type RecordProps = {
+    key: string;
+    data: any;
+    title: string;
+    form: React.ReactNode | null;
+    savePath: string;
+    onInsert?: (record: any, key: string) => any;
+    onUpdate?: (record: any, key: string) => any;
+    onDelete?: (record: any, key: string) => boolean;
+    onFinally?: (action: string, record: any, key: string) => any;
+    genKey: (record: any) => string;
+};
+const Grid = (props: GridProps) => {
+    return props.dataArray
+        ? <GridArray {...props} />
+        : <GridDatabase {...props} />;
 }
 
-function GridDatabase({
-                  columns = null,
-                  format = {},
-                  dataStoragePath = null,
-                  Header = null,
-                  HeaderAction = null,
-                  Footer = null,
-                  allowedActions = null,
-                  modalSize = "lg",
-                  setModalHeader = null,
-                  setPrimaryKey = null,
-                  onLoadRecord = null,
-                  onDisplayBefore = null,
-                  onInsertRecord = null,
-                  onUpdateRecord = null,
-                  onDeleteRecord = null,
-                  onFinallyRecord = null,
-                  onClick = null,
-                  Form = null,
-                  wrapClass = null,
-                  scroll = false,
-                  type = "table",
-                  showLoader =  false,
-                  groupBy = null,
-                  sticky= null,
-                  log= false
-              }: {
-    columns?: Array<{ label: string, key: string, sort?: boolean, onDisplay?: (value: string, record: any, key: string) => string }>,
-    format?: { [key: string]: string | ((data: string) => string) },
-    dataStoragePath?: string,
-    Header?: string | React.ReactNode,
-    HeaderAction?: string | React.ReactNode,
-    Footer?: string | React.ReactNode,
-    allowedActions ?: Array<"add" | "edit" | "delete">,
-    modalSize ?: string,
-    setModalHeader ?: (record ?: any) => string,
-    setPrimaryKey ?: (record: any) => string,
-    onLoadRecord?: (record: any, index, number) => any,
-    onDisplayBefore?: (records: any, setRecords: (records: any) => void, setLoader: (loader: boolean) => void) => void,
-    onInsertRecord?: (record: any, key: string) => any,
-    onUpdateRecord?: (record: any, key: string) => any,
-    onDeleteRecord?: (record: any, key: string) => any,
-    onFinallyRecord?: (action: string, record: any, key: string) => any,
-    onClick?: (record: any, key: string) => any,
-    Form?: any,
-    wrapClass?: string,
-    scroll?: boolean,
-    type?: "table" | "gallery",
-    showLoader?: boolean,
-    groupBy?: string | string[],
-    sticky?: "top" | "bottom",
-    log?: boolean
-}) {
-    const [records, setRecords] = useState(null);
+const GridDatabase = ({ dataStoragePath, ...props }: Omit<GridProps, 'dataArray'>) => {
+    const [records, setRecords] = useState<any>(null);
     const location = useLocation();
-    const dbStoragePath = dataStoragePath || (dataStoragePath === false ? null : trimSlash(location.pathname));
+    const dbStoragePath = dataStoragePath || trimSlash(location.pathname);
 
     db.useListener(dbStoragePath, setRecords);
 
-    return <GridArray
-                columns={columns}
-                format={format}
-                dataStoragePath={dataStoragePath}
-                dataArray={records}
-                Header={Header}
-                HeaderAction={HeaderAction}
-                Footer={Footer}
-                allowedActions={allowedActions}
-                modalSize={modalSize}
-                setModalHeader={setModalHeader}
-                setPrimaryKey={setPrimaryKey}
-                onLoadRecord={onLoadRecord}
-                onDisplayBefore={onDisplayBefore}
-                onInsertRecord={onInsertRecord}
-                onUpdateRecord={onUpdateRecord}
-                onDeleteRecord={onDeleteRecord}
-                onFinallyRecord={onFinallyRecord}
-                onClick={onClick}
-                Form={Form}
-                wrapClass={wrapClass}
-                scroll={scroll}
-                type={type}
-                showLoader={showLoader}
-                groupBy={groupBy}
-                sticky={sticky}
-                log={log}
-            />
-}
+    return <GridArray {...props} dataArray={records} />;
+};
 
-function GridArray({
-                  columns = null,
-                  format = {},
-                  dataStoragePath = null,
-                  dataArray = null,
-                  Header = null,
-                  HeaderAction = null,
-                  Footer = null,
-                  allowedActions = null,
-                  modalSize = "lg",
-                  setModalHeader = null,
-                  setPrimaryKey = null,
-                  onLoadRecord = null,
-                  onDisplayBefore = null,
-                  onInsertRecord = null,
-                  onUpdateRecord = null,
-                  onDeleteRecord = null,
-                  onFinallyRecord = null,
-                  onClick = null,
-                  Form = null,
-                  wrapClass = null,
-                  scroll = false,
-                  type = "table",
-                  showLoader =  false,
-                  groupBy = null,
-                  sticky = null,
-                  log = false
-              }: {
-    columns?: Array<{ label: string, key: string, sort?: boolean, onDisplay?: (value: string, record: any, key: string) => string }>,
-    format?: { [key: string]: string | ((data: string) => string) },
-    dataStoragePath?: string,
-    dataArray?: Array<{ [key: string]: any }>,
-    Header?: string | React.ReactNode,
-    HeaderAction?: string | React.ReactNode,
-    Footer?: string | React.ReactNode,
-    allowedActions ?: Array<"add" | "edit" | "delete">,
-    modalSize ?: string,
-    setModalHeader ?: (record ?: any) => string,
-    setPrimaryKey ?: (record: any) => string,
-    onLoadRecord?: (record: any, index, number) => any,
-    onDisplayBefore?: (records: any, setRecords: (records: any) => void, setLoader: (loader: boolean) => void) => void,
-    onInsertRecord?: (record: any, key: string) => any,
-    onUpdateRecord?: (record: any, key: string) => any,
-    onDeleteRecord?: (record: any, key: string) => any,
-    onFinallyRecord?: (action: string, record: any, key: string) => any,
-    onClick?: (record: any, key: string) => any,
-    Form?: any,
-    wrapClass?: string,
-    scroll?: boolean,
-    type?: "table" | "gallery",
-    showLoader?: boolean,
-    groupBy?: string | string[],
-    sticky?: "top" | "bottom",
-    log?: boolean
-}) {
+const GridArray = ({
+                       columns          = null,
+                       format           = {},
+                       dataStoragePath  = null,
+                       dataArray        = null,
+                       Header           = null,
+                       HeaderAction     = null,
+                       Footer           = null,
+                       allowedActions   = null,
+                       modalSize        = "lg",
+                       setModalHeader   = null,
+                       setPrimaryKey    = null,
+                       onLoadRecord     = null,
+                       onDisplayBefore  = null,
+                       onInsertRecord   = null,
+                       onUpdateRecord   = null,
+                       onDeleteRecord   = null,
+                       onFinallyRecord  = null,
+                       onClick          = null,
+                       Form             = null,
+                       scroll           = false,
+                       type             = "table",
+                       showLoader       =  false,
+                       groupBy          = null,
+                       sticky           = null,
+                       log              = false,
+                       wrapClass        = ""
+}: GridProps) => {
     const theme = useTheme();
 
     const [loader, setLoader] = useState(false);
     const [records, setRecords] = useState(null);
-    const [record, setRecord] = useState({ id: null, data: {} });
+    const [record, setRecord] = useState<RecordProps>(null);
     const [isModalOpen, setIsModalOpen] = useState(false);
     const refDomElem = useRef({});
     const location = useLocation();
@@ -297,28 +128,30 @@ function GridArray({
 
     if (!columns && records) {
         columns = (Form && typeof Form !== 'function'
-            ? extractComponentProps(Form, (child) => {
-                return { label: converter.toCamel(child.props.name, " "), key: child.props.name, sort: true}
-            })
-            : Object.keys(Object.values(records)[0] || {}).reduce((acc, key) => {
-                const value = Object.values(records)[0][key];
-                if (React.isValidElement(value) || typeof value !== 'object' || value === null || Array.isArray(value)) {
-                    acc.push({ label: converter.toCamel(key, " "), key: key, sort: true });
-                }
-                return acc;
-            }, [])
-        );
+                ? extractComponentProps(Form, (child) => {
+                    return { label: converter.toCamel(child.props.name, " "), key: child.props.name, sort: true };
+                })
+                : Object.keys(Object.values(records)[0] as { [key: string]: any }  || {}).reduce((acc: Column[], key) => {
+                    const value = Object.values(records)[0][key] as any;
+                    if (React.isValidElement(value) || typeof value !== 'object' || value === null || Array.isArray(value)) {
+                        acc.push({ label: converter.toCamel(key, " "), key: key, sort: true });
+                    }
+                    return acc;
+                }, [])
+        ) as Column[]
     }
 
     const columnsFunc = (columns || []).reduce((acc, column) => {
+        const formatKey = format?.[column.key];
+
         if (column?.onDisplay) {
             acc[column.key] = column.onDisplay;
-        } else if (format[column.key] && typeof format[column.key] === "function") {
-            acc[column.key] = format[column.key];
-        } else if (format[column.key] && typeof format[column.key] === "string") {
-            const convFunc = (converter[format[column.key]]
-                    ? format[column.key]
-                    : `to${ucfirst(format[column.key])}`
+        } else if (formatKey && typeof formatKey === "function") {
+            acc[column.key] = formatKey;
+        } else if (formatKey && typeof formatKey === "string") {
+            const convFunc = (converter[formatKey]
+                    ? formatKey
+                    : `to${ucfirst(formatKey)}`
             )
 
             acc[column.key] = ({value}) => (converter[convFunc] && converter[convFunc](value)) || value;
@@ -413,7 +246,7 @@ function GridArray({
             console.error("Save Failed: Key not generated correctly.");
             return;
         }
-        db.set(`${dbStoragePath}/${key}${record.savePath}`, data)
+        dbStoragePath && db.set(`${dbStoragePath}/${key}${record.savePath}`, data)
             .then(() => {
                 if (!record.key && key && refDomElem[key]) {
                     refDomElem[key].click();
@@ -426,7 +259,7 @@ function GridArray({
             .finally(() => {
                 closeModal();
 
-                const action = record.key ? "update" : "insert";
+                const action = record.key ? "update" : "create";
                 log && setLog(dbStoragePath, action, record.data, record.key);
                 return record.onFinally && record.onFinally(action, record.data, record.key);
             });
@@ -437,7 +270,7 @@ function GridArray({
             return;
         }
 
-        db.remove(`${dbStoragePath}/${record.key}`)
+        dbStoragePath && db.remove(`${dbStoragePath}/${record.key}`)
             .then(() => {
                 console.log('Element deleted successfully from Firebase');
             })
@@ -484,16 +317,15 @@ function GridArray({
                     headerClass={theme.Grid.Gallery.headerClass}
                     bodyClass={theme.Grid.Gallery.bodyClass}
                     footerClass={theme.Grid.Gallery.footerClass}
-                    selectedClass={!actionEdit && theme.Grid.Gallery.selectedClass}
+                    selectedClass={!actionEdit && (theme.Grid.Gallery.selectedClass)}
                     gutterSize={theme.Grid.Gallery.gutterSize}
                     rowCols={theme.Grid.Gallery.rowCols}
                     groupBy={groupBy}
-
                 />;
             case "table":
             default:
                 return <Table
-                    header={columns}
+                    header={columns || []}
                     body={body}
                     onClick={(onClick || actionEdit) && handleClick }
                     wrapClass={theme.Grid.Table.wrapperClass}
@@ -502,7 +334,7 @@ function GridArray({
                     headerClass={theme.Grid.Table.headerClass}
                     bodyClass={theme.Grid.Table.bodyClass}
                     footerClass={theme.Grid.Table.footerClass}
-                    selectedClass={!actionEdit && theme.Grid.Table.selectedClass}
+                    selectedClass={!actionEdit && (theme.Grid.Table.selectedClass)}
                 />;
         }
     }
